Run actual Python code from embedded interpreter
The embedded Python interpreter now contains enough code to
* show it's running and can import stuff, both
  - internally (from a string a la `python -c "..."`)
  - and from a file on the SDcard (like `python that_file.py`)
* log strings into the Android debug log (see `adb logcat`)
  via a function defined in C and bound into Python.

This version of the code assumes that the Python standard libraries
are available from `/sdcard/mypython`. In a proper implementation,
the app that instantiates the interpreter should hand down the
actual `getFilesDir()` (pointing to `/data/...` on internal
storage) or `getExternalStorageDirectory` (pointing to whatever
implements the `/sdcard` dir usually).