Modifies jh helper to always attach thread to JVM
JNI threading remains mysterious. We need to make sure that
each native thread created in Python gets its own thread
pointer to the JVM and also detaches when it stops.
Currently we simply attach the thread whenever we perform any
set of JNI actions in a jnihelper function. I think the overhead can
be neglected. So far we don't detach. FIX this!