Removes SensibilityApplication.java (static context)
SensibilityApplication.java was used to statically store a reference
to the app context.
This lead to problems when accessing it from Java code, called from native
code running in App Services running with different processes than the
main activity.
Now, each services pass their own context to the native code and
which passes it back to Java code that requires access to a
context (e.g. to call Sensors or send messages to the UI)