Adds basic notification C-Python extension
Simpler than simple[1] notification C-Python extension to print message to
Android UI notification drawer.

The notification has no click handler, so if the message is too long for
the screen it gets truncated by the system and there is no way to see the
entire message.

Also the notification id is currently hardcoded into the Java function,
therefore new calls to the function will update an existing notification
in the drawer.

[1] https://developer.android.com/guide/topics/ui/notifiers/notifications.html