Add vibration support
This commit adds the Java/JNI/CPython parts of a function that vibrates
the device for the given number of seconds. (Yes, *seconds*! We convert
to milliseconds internally, as required by the Android API.)

For this, add
* the `VIBRATE` permission to the app manifest
* the `vibrate` method to `OutputService.java`. This is where the
  `Vibrator.vibrate` method of Android is eventually called.
* the `androidlog_vibrate` Python/C/JNI/Java glue to `jni/outputs.c`
* the appropriate mapping to a Python function named `vibrate`
* the cached function `cached_output_vibrate` to `jni/snakei.c`
...and a few bits and pieces in the various header files.

See also SensibilityTestbed/repy_v2@8241b3593650a15d90e398838796f22e744f0c36
which adds the `androidlog.androidlog_vibrate` function to the RepyV2
namespace as `vibrate`.