Fix SensibilityTestbed/repy_v2#6, NaN not in JSON
Apart from a small debug addition to show which sensor number
is being stopped, and a few refactored variable names, this
commit adds checks for NaN's ("not a number") that some sensor
properties can contain. NaN's are valid `float`s, but not
representable in JSON. If we encounter a NaN in one of the
sensor properties that we query, we just omit this property from
the sensor property dict.

Also record other JSON errors that we may encounter during querying
the sensor properties. (This isn't really cleanly implemented right
now, but I've left an `@todo` in the method docstring).

Implementation note: I chose to go for a relatively generic way
(`Map`and loop) to check for NaN's. I found this cleaner than an
equally long chain of `if (! Float.isNaN(...)) {...}`. Also, it
sketches how the outer structure of looping over all of the sensors
and their property getters could look like.
Thanks @fmetzger for the helpful hints!