Fixes missing raw resources (seattle_android.zip)
If the user chooses to install seattle from a pre-packaged installer,
that installer has to be in res/raw/seattle_android.zip.
Prior to this commit a missing installer gave a "cannot find symbol"
error even if the user intended to install a downloaded installer.

This commit fixes the error by checking whether the pre-packaged
installer exists, and disables the install from zip button otherwise.