Fix raw resource install overrides download install
It is nice to be able to download an installer instead of using
the pre-packaged one. It does not help much if the install routine
overrides the downloaded files with the pre-packaged though.

This commit separates the download installer and move installer
from raw resource from the actual seattle installer (python routine).