Replaces path attrs with relative path class vars
SensibilityActivity now defines constants for relative paths
(Python, Seattle, Vessel) that need to be prefixed by
the data directory path.
We do this because we want to access those constants from other
components (PythonInterpreterService) and don't want to rely
on them being initialized by an instance of SensibilityActivity.

This replaces the full path attribute initializer and adds
a static data directory path get function that requires a
context.

This is not ideal yet. Maybe this would be a Use Case for
an Application Class?