Partially fix SensibilityTestbed/repy_v2#4, handle `SecurityException`s
`SecurityException`s are raised when the app tries to access
a sensor which it isn't allowed (yet) to access. For example,
the user might not have seen the query for allowance, or they
deliberately turned off a previously granted permission through
the Android "app settings" GUI.

Anyway, this commit wraps the problematic calls in `MiscInfoService`
in `try/catch` blocks and just returns `null` for the sensors
values (or sub-values in the case of sensors returning dicts)
that the app isn't allowed to access.

The other sensor classes (location, media, output, sensor)
still need to be checked!

(In an ideal world, and if we only had to support higher API
levels, we could ask the user again and again whether they
want to grant the permission this time, and even provde them
reasons for why we think we should be granted acces. For now
however, it seems that hiding the `SecurityException` from
sandbox code is a relatively safe and sane thing to do.)