Adds IP(v4) address fragment
This fragment shows the current local WiFi IP address of the
device. This is useful in case an experimenter wants to "contact"
(seash) an donated device, which requires to given an IP address.

The IP address is updated each time the user changes to the fragment.
The fragment also shows a date time when it was last updated.