Dedicates service process to nodemanager (WIP)
- Renames PythonInterpreterService0 to PythonNodemanagerService
and the reserved process from proc0 to nmmain.

- NM service overrides static startService which does not look for
an idle service (as opposed to its super) but starts itself
if it is not already running.

- Adds static function to find out if a specific service process
is running

- Checks if nmmain is running in Auto/Develop installation and
starts as required

Note: In the auto installation procedure the nodemanager seems to
start but crashes right afterwards.
Maybe the installation process has not finished?