Changes isSeattleInstalled to require line in log
Previously, we assessed whether seattle was installed by looking
for the existence of the service vessel. Since the service vessel
is created before the installation has finished, the nodemanager
was sometimes started to early.

Now we check if seattle_repy contains installerstdout.log and look
for a line "seattle has been installed". This check should be
reliable but might be more costly than needed.

We also give the auto installer more time to install before we
start the nodemanager or abort with "not installed in timely manner".