Disables auto install btn if all up and running
The auto installation button is enabled if any of
python is not installed, or custom installer is not
downloaded or not installed, or nodemanager is not running.

If everything is up and running, we can disable this button.

Note: the buttons in the develop tab are never disabled, because
they can used to forcefully start/stop/re-install independently
of the current state of the system.