Adds uri data C-Python extension (heavy WIP)
Adds functionality to receive data in Python that was sent when
someone clicked a special URL. The idea is:

1. Define an intent-filter in the manifest that starts app's main
activity if the user clicks on a link in a special format,
e.g: sensi://send.data

2. In the app's main activity's onStart we check if the app intent
came from a user clicking on a link. If yes, we display a toast
message and forward the intent to a broadcast receiver which
can be accessed from a service.

3. When running a native process we initialize a DataService
singleton using the usual cumbersome JNI calls.
That data service programmatically registers a broadcast receiver
with an intent filter for intents coming from (2).
If it receives an intent with URI and optionally query parameter
encoded data, it stores it to an attribute of DataService.

Note: Unfortunately we cannot directly listen for intents from (1),
they have to be delivered to an activity.

4. The DataService's attribute(s) containing the most recent data
received from a link-click-intent (or all data received so far)
can be accessed via C-Python Extensions: i.e.
androiddata.get_data() and androiddata.get_all_data()

DOES NOT WORK YET!!